fix(gen/convert_image): issues with image handling
Fixed issue which threw 500 error when uploading a file. This was because the `TempFile`
was being saved without the file extension and the image conversion could not decide what
the file type was.

Also fixed oversight where, if a newly generated file name was found to exist already,
a new file name would be generated but never be returned. This meant the old, existing
file name would be baked into the returned url in the response.